Changed "seamless" option setting to use object consisting of enabled: BOOL and honestPageCount: BOOL.
The Honest Page Count setting is disabled by default when seamless is enabled. This is to preserve familiar functionality in cases like the gotopage and lightbox examples. When honestPageCount is disabled, the USEFUL page count will be returned. The first set of pages before repeating elements is returned. This will cause some elements to be cut off in cases where visible % total elements != 0, but will allow consistent page transitions.